# httpapi/openapi.yaml
openapi: "3.1.0"
info:
  title: Pit Officer REST API
  version: "1"
  description: >
    Control-plane REST API for the Pit Officer service. The public surface is
    mounted at `/api/v1`; a mirror at `/app/api/v1` is reserved for the operator
    dashboard and carries the same routes except the route manifest, which is
    published only on `/api/v1`. Control endpoints answer `application/json`;
    endpoints that hand back a file answer with that file's own media type.
    A malformed request body, or one that violates the schema or a parameter
    constraint, answers RFC 9457 `application/problem+json` - `400` for broken
    syntax, `422` for a constraint violation. Every other error answers the
    envelope `{"error":{"code":"...","message":"..."}}`.
    Money and size values
    are exact decimal strings - never IEEE 754 floats.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
    description: Public REST API
...
paths:
...
  /mcp-access:
    get:
      operationId: listMcpAccess
      summary: List the MCP command catalogue and per-command enabled state
      description: >
        Returns every MCP command Pit Officer exposes - both the implemented
        read-only tools and the designed-but-not-yet-built mutating ones - with
        its metadata and current effective enabled state. A command with no
        stored override resolves to its catalogue default (read-only commands
        on, protective commands off).
      tags: [mcp-access]
      responses:
        "200":
          description: MCP command catalogue with effective enabled state.
          content:
            application/json:
              schema:
                type: object
                required: [commands]
                properties:
                  commands:
                    type: array
                    items:
                      $ref: "#/components/schemas/McpCommand"
        "500":
          $ref: "#/components/responses/InternalError"
  /mcp-access/{command}:
    parameters:
      - name: command
        in: path
        required: true
        description: MCP command name (must be in the catalogue).
        schema:
          type: string
    put:
      operationId: setMcpAccess
      summary: Enable or disable one MCP command
      description: >
        Persists the enabled state for one catalogue command. A disabled command
        stays callable over MCP but returns a non-error "disabled in panel"
        notice instead of running. Confirmation-on-enable for protective commands
        is a UI concern; the backend just persists.
      tags: [mcp-access]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [enabled]
              properties:
                enabled:
                  type: boolean
                  description: Whether the command is enabled.
      responses:
        "200":
          description: The updated command with its persisted enabled state.
          content:
            application/json:
              schema:
                type: object
                required: [command]
                properties:
                  command:
                    $ref: "#/components/schemas/McpCommand"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # User settings
  # ---------------------------------------------------------------------------
...
components:
...
  responses:
    MalformedRequestError:
      description: Malformed HTTP request or JSON syntax.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ValidationError:
      description: Syntactically valid request failed schema or parameter validation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, errors]
      properties:
        type:
          type: string
          description: >
            RFC 9457 problem type URI. Officer uses `about:blank` for generic
            HTTP errors.
          enum: ["about:blank"]
        title:
          type: string
          description: Canonical HTTP status title.
        status:
          type: integer
          enum: [400, 422]
        detail:
          type: string
          description: Human-readable explanation for this occurrence.
        errors:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ProblemError"
    ProblemError:
      type: object
      additionalProperties: false
      required: [code, pointer]
      properties:
        code:
          type: string
          description: Stable machine-readable error category.
        constraint:
          type: string
          description: Violated request constraint, when known.
        pointer:
          type: string
          description: JSON Pointer to the invalid member; empty means the request root.
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: >
                Machine-readable code: `validation`, `not_found`,
                `account_missing`, `conflict`, `currency_change_blocked`, `has_dependents`,
                `terminal_order`, `execution_report_required`, `precondition`,
                `too_large`, `forbidden`, `reserved_group`,
                `engine_restarting`, `signing`, `upstream`, `not_implemented`,
                or `internal`.
              example: not_found
            message:
              type: string
              example: account not found
...
    McpCommand:
      type: object
      required:
        [name, title, agentDescription, mutating, protective, implemented, enabled]
      properties:
        name:
          type: string
          description: Wire name of the MCP tool.
        title:
          type: string
          description: Human-readable label for the operator panel.
        agentDescription:
          type: string
          description: One-line description of what an AI agent does with the command.
        mutating:
          type: boolean
          description: Whether the command changes control-plane or engine state.
        protective:
          type: boolean
          description: >
            Whether the command guards capital or risk posture, so the panel can
            require confirmation before enabling it.
        implemented:
          type: boolean
          description: Whether a handler backs the command yet.
        enabled:
          type: boolean
          description: >
            Effective enabled state: the stored override when present, otherwise
            the catalogue default.

    # Signing keys and approval token
...